+++ oid_validation/setup.py
@@ -2,10 +2,15 @@ from setuptools import setup, find_packages
 
 
 setup(
-    name='pulp-repoauth',
+    name='pulp-oid_validation',
     version='2.7.0a2',
     license='GPLv2+',
     packages=find_packages(exclude=['test']),
     author='Pulp Team',
     author_email='pulp-list@redhat.com',
+    entry_points={
+        'pulp_content_authenticators': [
+            'oid_validation=pulp.oid_validation.oid_validation:authenticate'
+        ]
+    },
 )
