+++ oid_validation/test/data/cert.crt
@@ -1,0 +1,20 @@
+-----BEGIN CERTIFICATE-----
+MIIDOjCCAiICAQowDQYJKoZIhvcNAQEFBQAwYTELMAkGA1UEBhMCVVMxCzAJBgNV
+BAgMAk5KMRIwEAYDVQQHDAlNaWNrbGV0b24xEDAOBgNVBAoMB1JlZCBIYXQxDTAL
+BgNVBAsMBFB1bHAxEDAOBgNVBAMMB3B1bHAtY2EwHhcNMTIwMzI4MjExMzQ5WhcN
+MTYwMzI3MjExMzQ5WjBlMQswCQYDVQQGEwJVUzELMAkGA1UECAwCTkoxEjAQBgNV
+BAcMCU1pY2tsZXRvbjEQMA4GA1UECgwHUmVkIEhhdDENMAsGA1UECwwEUHVscDEU
+MBIGA1UEAwwLcHVscC1zZXJ2ZXIwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEK
+AoIBAQDOVa5Iolo3FqaPB0B/hFSgM0uX1PYNXyWzgVneXdwYhAR+10v5gc97aefd
+VkYSodVbIRBwi+SC7QSuW1gpUdsY/fyVrpmSB4Vyszgn+7VPxcLhlY3bYHKu6sNh
+WBQ7+39Y5s21sRrmR1PboC9ydS7XeRLpefycadYKj8CMiN/UMI6F/8uKDspyOZ1m
+08njc8kzZUyR3ITpHngnRI+UVtTqU9cPhJwBZuUnXHjM39Obiv8jrmrjgm1hxGWW
+vAYaGkd5hHER0Nyy/WeSnWyxUSPeLSWKMPwRkPzguOFZTa0etB5+Qg3BjN47ZNiV
+Qw6QuToiTtJpPszgI+rwZ0cxBR4FAgMBAAEwDQYJKoZIhvcNAQEFBQADggEBACaq
+hUwpiEQrdhbAM2502/5b+tJpQ6W+l2+rbWfRW/rO3Bt6nLZv54bzjCpw+vGf9W0t
+glU2cQ1NcT3SB0miFlmvhisWX9azlz5XYa8kH0hd/Hp+LPUNsOV42E+A+oGg7OAD
+fciuQXQjiH6xCpsvLa8HW/QEoWyW3tpD4KV1FwV0lFP42t2T7liScSTkcLhYWhil
+C9ctawA/HhAX0EkE9VxkLiYhyifb3jCBQDQ8KPILJ2olGCAoGrZe83AeGo96L4wS
+b23vr/F8g6Eq20qCcxtvtmECIpa+5ZYJQXkxpEMvSWvfz24ATCvTGnSPIgxLiJbs
+vqtsIu26ic0SgDtV9Uw=
+-----END CERTIFICATE-----
