+++ oid_validation/test/data/invalid_ca.crt
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDpTCCAo2gAwIBAgIJAKkrGxLzlBlLMA0GCSqGSIb3DQEBBQUAMGkxCzAJBgNV
+BAYTAlVTMQswCQYDVQQIDAJOSjESMBAGA1UEBwwJTWlja2xldG9uMRAwDgYDVQQK
+DAdSZWQgSGF0MQ0wCwYDVQQLDARQdWxwMRgwFgYDVQQDDA9wdWxwLWludmFsaWQt
+Y2EwHhcNMTIwNDE2MTM0NTIxWhcNMjUxMjI0MTM0NTIxWjBpMQswCQYDVQQGEwJV
+UzELMAkGA1UECAwCTkoxEjAQBgNVBAcMCU1pY2tsZXRvbjEQMA4GA1UECgwHUmVk
+IEhhdDENMAsGA1UECwwEUHVscDEYMBYGA1UEAwwPcHVscC1pbnZhbGlkLWNhMIIB
+IjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAqdGWUTKTf74EsxrT7+XqDh3F
+muGiJIRs4BAZPBMCenxxag/FRwtWeR/Z8xovLvlpivOvJiYxDlmQROMH9AbzCT94
+qChgre+weel81bWR6gc4fOxUTap4cHA7nwvP5PQP6oUlcn0ZH0fyN/YAj1c0d7Bv
+UHyne+C1fxcYIjgouERHWo9P0xv7jpnn2nUxs40B4QDTK+g0Zu4h/KGiEVr6IK+/
+ZRO8KCih9UdzVYKG8AjeY0d0LP3eSuX8srcM2RFwFcXh3gpt5aWgcUQR3OjeSHBR
+ho/yDXxm52tsD7J5QXvijvu7ILuE7yUVVkCxtOu7POtEKf/aDjQiMh194MZ0gQID
+AQABo1AwTjAdBgNVHQ4EFgQUtlwrMUV5vmkU/UJq3xBGCGmQmIMwHwYDVR0jBBgw
+FoAUtlwrMUV5vmkU/UJq3xBGCGmQmIMwDAYDVR0TBAUwAwEB/zANBgkqhkiG9w0B
+AQUFAAOCAQEABlL6CBQBEqGreP8hvc8mm9YWb4SOZdCreEuewMYdV0tdIiS6rjg6
+xoQhAzWmvBVxd0kpm33TAP9mqD9oExrH7WTc+QCRoihW7EcoK5utXAeU8oiuFSSh
+zZUBkCBQkDX7QF0twLorfKxfNEuNuUj1anGHEjESadQV++dNl9yvM82JcpqgAuoj
+rdAaDQrVVRVpCe5ClJqWJROziEEGj10nsTskjuqXChaslJ2O0iYm6ZPZcmDXOOEj
+yF0ir5JjvZQ6zZdo/D+wSdfK1TLl5hjpzFTlElOeOC5XM13pgUfIF3nWeIKJEUyJ
+YSMf0fu6BrpTgoyet283Ek9qg8NqKtMv1A==
+-----END CERTIFICATE-----
