+++ oid_validation/test/data/test-override-pulp.conf
@@ -1,0 +1,32 @@
+[server]
+server_name: localhost
+relative_url: /pulp/repos
+default_login: admin
+default_password: admin
+storage_dir: /tmp/pulp
+
+[logs]
+config: /etc/pulp/logging.cfg
+
+[rhn]
+threads: 25
+
+[auditing]
+events_file: /tmp/pulp/events.log
+lifetime: 0
+
+[repos]
+cert_location: /tmp/etc/pki/content
+global_cert_location: /tmp/etc/pki/content
+protected_repo_listing_file: /tmp/etc/pki/content/pulp-protected-repos
+
+[database]
+name: pulp_unittest_rpm_support
+
+[security]
+oauth_key: some-key
+oauth_secret: some-secret
+
+# Merged with the following from repo_auth.conf
+[main]
+enabled: false
