+++ oid_validation/test/data/test-override-repoauth.conf
@@ -1,0 +1,19 @@
+[main]
+enabled: false
+log_failed_cert: true
+log_failed_cert_verbose: false
+max_num_certs_in_chain: 100
+# If this is true, the client certificate will be verified by Pulp against the per-repo certificate
+# authorities. If it is false, client certificates will not be checked for signature or expiration.
+# If you don't need per-repo CAs, it is recommended to set this to false and use your web server to
+# check the client certificates against a trusted CA back. Setting this to true will have a negative
+# performance impact, so don't do that unless you need per-repo CAs. It is true by default to
+# maintain backwards compatibility.
+# verify_ssl: true
+
+[repos]
+cert_location: /tmp/test_oid_validation/
+global_cert_location: /tmp/test_oid_validation/
+protected_repo_listing_file: /tmp/test-protected-repo-listing
+
+
