+++ oid_validation/test/data/valid_ca.crt
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDlTCCAn2gAwIBAgIJAP+15ciSiPzeMA0GCSqGSIb3DQEBBQUAMGExCzAJBgNV
+BAYTAlVTMQswCQYDVQQIDAJOSjESMBAGA1UEBwwJTWlja2xldG9uMRAwDgYDVQQK
+DAdSZWQgSGF0MQ0wCwYDVQQLDARQdWxwMRAwDgYDVQQDDAdwdWxwLWNhMB4XDTEy
+MDMyODIxMDczMVoXDTE2MDMyNzIxMDczMVowYTELMAkGA1UEBhMCVVMxCzAJBgNV
+BAgMAk5KMRIwEAYDVQQHDAlNaWNrbGV0b24xEDAOBgNVBAoMB1JlZCBIYXQxDTAL
+BgNVBAsMBFB1bHAxEDAOBgNVBAMMB3B1bHAtY2EwggEiMA0GCSqGSIb3DQEBAQUA
+A4IBDwAwggEKAoIBAQC4fIQQ3Rx9y0x0/GdZeD+xypzh8ytQBX7nLavBp0sslF5G
+xWyO2kBq4NY3LtQOxPamDpWQck4eBps/6ajMTeC09hKHxXCHcD2Ivx/kO72bev6z
+gSVKUSJnu1AF4fIlQl/nxodKz6XwSR4C9+GfK9jNP6NUFQKGsQ3/FHDCvEUkdkG9
+SDXfxRRq6FTAl+Owt7JXIGzosbU6d8mZoqpRx6AZ+kA7Bas2lBG0km785SMyxjVB
+CabS09M0gPzHPBfrucIgvIyw82VJY25UWsa1UkQyfJ/nsuj1WhWJa6XvgtsZfTm4
+56mFcFgu3hJ8I0SLvnsTZVoIZoUKMj/h5FwhcBs3AgMBAAGjUDBOMB0GA1UdDgQW
+BBQUEhXReUDSCm2ZOrkbOPuFedhl2jAfBgNVHSMEGDAWgBQUEhXReUDSCm2ZOrkb
+OPuFedhl2jAMBgNVHRMEBTADAQH/MA0GCSqGSIb3DQEBBQUAA4IBAQABcUhYK9NY
+5ew5mlVIzEPgZDdWhDHRutottQb3UQu6c2kHwwJZdzplHaxn5Mo0fqt6JaYbqnbH
+goLCHyLpRLBL/Dj/7SFM8xKgkFkLfd4fDOINV6Gec3tIEFBDBG3wcYT5GcFQh+JD
+Qga6Lt57gLnkd9G8Y93It1T0pbFQhOh8U+g/jnPjCqFTqOMoAngVPZZTsg26cVh+
+12jOyy7kXrKEtLESaj56rFoRtX8QbwknYJ60W8t+n2ZqtCjyNN65RXTFK95TmWSZ
+hYMwmwaaUUDm4wSKS/4mmUKFCN8eaCGHKFN47hK66KZ9vRKK3VFOssQ5CIE8o2b7
+/Yb4KafFsHUY
+-----END CERTIFICATE-----
